reorganized commands into folders
