+++ src/main/java/frc/robot/RobotContainer.java
@@ -20,12 +20,12 @@ import edu.wpi.first.wpilibj2.command.WaitCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.Constants.ElevatorConstants.ElevatorPositions;
-import frc.robot.commands.ArmToAngleCmd;
-import frc.robot.commands.DefaultFunnelCommand;
-import frc.robot.commands.JamesHardenMovement;
-import frc.robot.commands.RunFunnelUntilDetection;
-import frc.robot.commands.SetElevatorLevel;
-import frc.robot.commands.SwerveJoystickCommand;
+import frc.robot.commands.DaleCommands.ArmToAngleCmd;
+import frc.robot.commands.ElevatorCommands.SetElevatorLevel;
+import frc.robot.commands.FunnelCommands.DefaultFunnelCommand;
+import frc.robot.commands.FunnelCommands.RunFunnelUntilDetection;
+import frc.robot.commands.SwerveCommands.JamesHardenMovement;
+import frc.robot.commands.SwerveCommands.SwerveJoystickCommand;
 import frc.robot.subsystems.ArmSubsystem;
 import frc.robot.subsystems.ElevatorSubsystem;
 import frc.robot.subsystems.FunnelSubsystem;
