+++ src/main/java/frc/robot/commandGroups/LoadAndPutUp.java
@@ -6,10 +6,10 @@ package frc.robot.commandGroups;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants.ElevatorConstants.ElevatorPositions;
-import frc.robot.commands.ElevatorIntakeLevel;
-import frc.robot.commands.RunFunnelUntilDetection;
-import frc.robot.commands.SetElevatorLevel;
 import frc.robot.commands.TransferPieceBetweenFunnelAndElevator;
+import frc.robot.commands.ElevatorCommands.ElevatorIntakeLevel;
+import frc.robot.commands.ElevatorCommands.SetElevatorLevel;
+import frc.robot.commands.FunnelCommands.RunFunnelUntilDetection;
 import frc.robot.subsystems.ElevatorSubsystem;
 import frc.robot.subsystems.FunnelSubsystem;
