+++ src/main/java/frc/robot/commands/ElevatorCommands/SetElevatorLevel.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.ElevatorCommands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.ElevatorConstants.ElevatorPositions;
