+++ src/main/java/frc/robot/commands/FunnelCommands/RunFunnelUntilDetection.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.FunnelCommands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.FunnelSubsystem;
