+++ src/main/java/frc/robot/commands/SwerveCommands/JamesHardenMovement.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.SwerveCommands;
 
 import static edu.wpi.first.units.Units.Meters;
