+++ src/main/java/frc/robot/commands/TootsieSlideCommands/TootsieSlideShooting.java
@@ -2,20 +2,19 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.TootsieSlideCommands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.ArmSubsystem;
+import frc.robot.subsystems.TootsieSlideSubsystem;
 
-/* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class ZeroArm extends Command {
-  /** Creates a new ZeroArm. */
-  ArmSubsystem arm;
+/** An example command that uses an example subsystem. */
+public class TootsieSlideShooting extends Command {
+  @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
+  private TootsieSlideSubsystem tootsieSlideSubsystem;
 
-  public ZeroArm(ArmSubsystem arm) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.arm = arm;
-    addRequirements(this.arm);
+  public TootsieSlideShooting(TootsieSlideSubsystem subsystem) {
+    tootsieSlideSubsystem = subsystem;
+    addRequirements(tootsieSlideSubsystem);
   }
 
   // Called when the command is initially scheduled.
@@ -25,20 +24,18 @@ public class ZeroArm extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    arm.moveMuyNegative();
+    tootsieSlideSubsystem.shootTootsie();
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    if (!interrupted) {
-      arm.zeroSensor();
-    }
+    tootsieSlideSubsystem.stopTootsie();
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return arm.checkCurrent();
+    return false;
   }
 }
